pinctrl: rockchip: fix non-hiword-mask registers on rk3288
On the rk3288 the pinctrl-related registers do not use the hiword-mask
paradigm (enabling the the corresponding bit for writing by setting
bit x+16 to 1). Instead these registers need common update operations.

In contrast the pull-settings residing in the pmu area on the rk3188 use
the hiword-mask paradigm and from what I heard, future SoCs are supposed
to use the hiword-mask in the complete pmu area, so this will hopefully
stay a one-off issue.

Signed-off-by: Heiko Stuebner <heiko@sntech.de>