UPSTREAM: ARM: dts: Turn on USB host vbus on rk3288-evb
There is no phy driver that works on the Rockchip board for either USB
host port yet.  For now just hardcode the vbus signal to be on all the
time which makes both the dwc2 host and the EHCI port work.

BUG=chrome-os-partner:30973
TEST=With series EHCI port works on evb

Change-Id: I0d645f6e132ae442cd20c1adb0e254c7fdee8e50
Signed-off-by: Doug Anderson <dianders@chromium.org>
Signed-off-by: Heiko Stuebner <heiko@sntech.de>
(cherry picked from git.kernel.org arm/arm-soc.git for-next
 commit 83a742b431501c1cf971983fa1911ae5d4b0440c)
Reviewed-on: https://chromium-review.googlesource.com/208240

Conflicts:
	arch/arm/boot/dts/rk3288-evb.dtsi