ARM: dts: Move the PMU interrupt pinctrl line to rk3288-evb common
The PMU interrupt pinctrl line was added to the rk3288-evb-act8846,
but it's the same line on both the ACT8846 version and the RK808
version.  This makes a lot of sense since they share the same SoC
daugherboard.  Move the pinctrl definition to the common file so we
can use it for the RK808 version.

BUG=chrome-os-partner:30167
TEST=Used by future change.

Change-Id: I7249b902f87a6e3c251a162ea9de5252c0f0cd3b
Signed-off-by: Doug Anderson <dianders@chromium.org>