thermal: rockchip: add driver for Rockchip thermal
Thermal is TS-ADC Controller module supports user-defined mode and automatic mode.

User-defined mode refers,TSADC all the control signals entirely by software
writing to register for direct control.

Automaic mode refers to the module automatically poll TSADC output,and the results
Were checked.

If you find that the temperature High in a period of time, an interrupt is generated
to the processor down-measures taken;if the temperature over a period of time High,
the resulting TSHUT gave CRU module,let it reset the entire chip, or via GPIO give PMIC.

Signed-off-by: zhaoyifeng <zyf@rock-chips.com>
Signed-off-by: Caesar Wang <caesar.wang@rock-chips.com>